requirements.txt
streamlit
requestsstream